タスク11の編集

+++ src/main/java/jp/co/seattle/library/controller/DeleteBookController.java
@@ -13,39 +13,51 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import jp.co.seattle.library.service.BooksService;
+import jp.co.seattle.library.service.RentService;
 
 /**
  * 削除コントローラー
  */
-@Controller //APIの入り口
+@Controller // APIの入り口
 public class DeleteBookController {
-    final static Logger logger = LoggerFactory.getLogger(DeleteBookController.class);
-    //booksService book = new  booksService();
-    @Autowired
-    private BooksService booksService;
+	final static Logger logger = LoggerFactory.getLogger(DeleteBookController.class);
+	// booksService book = new booksService();
+	@Autowired
+	private BooksService booksService;
 
-    /**
-     * 対象書籍を削除する
-     *
-     * @param locale ロケール情報
-     * @param bookId 書籍ID
-     * @param model モデル情報
-     * @return 遷移先画面名
-     */
-    @Transactional
-    @RequestMapping(value = "/deleteBook", method = RequestMethod.POST)
-    public String deleteBook(
-            Locale locale,
-            @RequestParam("bookId") Integer bookId,
-            
-            Model model) {
-        logger.info("Welcome delete! The client locale is {}.", locale);
-        booksService.deleteBook(bookId);
-        model.addAttribute("bookList", booksService.getBookList());
+	@Autowired
+	private RentService rentservice;
 
+	/**
+	 * 対象書籍を削除する
+	 *
+	 * @param locale ロケール情報
+	 * @param bookId 書籍ID
+	 * @param model  モデル情報
+	 * @return 遷移先画面名
+	 */
+	@Transactional
+	@RequestMapping(value = "/deleteBook", method = RequestMethod.POST)
+	public String deleteBook(Locale locale, @RequestParam("bookId") Integer bookId,
 
-        return "home";
+			Model model) {
+		logger.info("Welcome delete! The client locale is {}.", locale);
 
-    }
+		int rentalId = rentservice.rentBook(bookId);
+
+		if (rentalId == 0) {
+			booksService.deleteBook(bookId);
+			model.addAttribute("bookList", booksService.getBookList());
+
+		} else {
+			rentservice.returnBook(bookId);
+			model.addAttribute("error", "この本は貸し出し中のため、削除できません。");
+			model.addAttribute("bookDetailsInfo", booksService.getBookInfo(bookId));
+			return "details";
+		}
+
+		return "home";
+
+	}
 
 }
